Added save functionality to categories
